==> itas_analysis/notebooks/1_scrap_itas_web2010-2017_2.ipynb <==
import urllib
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np

itas2010 = 'http://www.itas2010.iitp.ru/search.html'
itas2011 = 'http://www.itas2011.iitp.ru/search.html'
itas2012 = 'http://www.itas2012.iitp.ru/search.html'
itas2013 = 'http://itas2013.iitp.ru/ru/search.html'
itas2014 = 'http://itas2014.iitp.ru/ru/search.html'
itas2015 = 'http://itas2015.iitp.ru/ru/search.html'
itas2016 = 'http://itas2016.iitp.ru/ru/search.html'
itas2017 = 'http://itas2017.iitp.ru/papers/'

def load_authors_titles(url, year):
    import urllib
    from bs4 import BeautifulSoup
    
    response = urllib.request.urlopen(url)
    if response.status == 200:
        soup = BeautifulSoup(response, 'html.parser')
        
        titles = soup.find_all("a", {"class": "PaperTitle"})
        authors = soup.find_all("span", {"class": "PaperAuthors"})
        
        authors_list = []
        titles_list = []
        
        for auth, title in zip(authors, titles):
            authors_list.append(auth.get_text())
            titles_list.append(title.get_text())
            
        years = [year]*len(authors_list)
        
        return authors_list, titles_list, years
            
    else:
        print(response)
        return -1, -1, -1
    
    
def get_itas2017():    
    def get_author(text):
        try:
            all_authors, other = text.split('"other_authors"')
            other_authors, other = other.split('"first_author"')
            first_author, other = other.split('"platform"')
            other_authors, pdf= other_authors.split('"pdf"')
            other = bytes(other_authors, 'utf-8').decode('unicode_escape')
            first = bytes(first_author, 'utf-8').decode('unicode_escape')

            all_authors = []
            for f in first.split('"'):
                if len(f) > 5:
                    all_authors.append(f)

            for o in other.split('"'):
                if len(o) > 7:
                    all_authors.append(o)
            return all_authors
        except:
            return []
    data = []
    authors=[]
    response = urllib.request.urlopen(itas2017)
    soup = BeautifulSoup(response, 'html.parser')
    papers = soup.text.split('"id"')
    
    for idx,p in enumerate(papers):
        try:
            info = p.split('"abstract"')[1].split('"title"')[1]
            info2 = info.split('"all_authors"')
            title = bytes(info2[0], 'utf-8').decode('unicode_escape')
            for t in title.split('"'):
                if len(t)> 7:
                    main_title = t
            data.append(main_title)#(title)
            first_author = get_author(info2[1])
            authors.append(first_author)
        except:
            pass
        
    #authors[40], authors[42], authors[77], authors[102]
    #([], [], [], [])
    
    authors[40] = ['Анастасия Жарикова']
    authors[42] = ['Igor Zhilin', 'Victor Zyablov']
    authors[77] = ['Дарья Николаева', 'Софья Гарушянц']
    authors[102] = ['Nadezhda Terekhanova']
    
    _authors = []
    for auth in authors:
        _authors.append(', '.join(auth))
    
    return data, _authors


def process_authors(auth):
    a = []
    auth = auth.strip()
    for x in auth.split(','):
        a.append(x.strip())
    return a

## Store info in DataFrame, combine all years

authors, titles, years = load_authors_titles(itas2010, 2010)

itas_all = pd.DataFrame(data={'authors': authors,
                          'titles':titles,
                          'year':years}, 
                    columns=[['authors', 'titles', 'year']])

for url,year in zip([itas2011, itas2012, itas2013, itas2014, itas2015, itas2016],
                   [2011, 2012, 2013, 2014, 2015, 2016]):

    authors, titles, years = load_authors_titles(url, year)

    itas = pd.DataFrame(data={'authors': authors,
                              'titles':titles,
                              'year':years}, 
                        columns=[['authors', 'titles', 'year']])

    itas_all = pd.concat([itas_all, itas])

titles, authors = get_itas2017()

years = [2017]*113
itas = pd.DataFrame(data={'authors': authors,
                              'titles':titles,
                              'year':years}, 
                        columns=[['authors', 'titles', 'year']])
itas_all = pd.concat([itas_all, itas])

itas_all.drop_duplicates(inplace=True)

authors = []
titles = []
years = []
for row in itas_all.iterrows():
    #print(type(row))
    #print(row[1].authors)
    article_authors = process_authors(row[1].authors)
    for auth in article_authors:
        authors.append(auth)
        titles.append(row[1].titles)
        years.append(row[1].year)

itas = pd.DataFrame(data={'author':authors,
                   'title':titles,
                   'year':years}, columns=[['author', 'title', 'year']])

itas.iloc[[678, 1114, 1199, 1964]].values

# itas[itas.title == 'Количественное исследование активности центральных зрительных нейронов лягушки в зависимости от угловой скорости движения внешнего раздражителя ']
# Третий автор Владимир Антонинович Бастаков
itas = itas.set_value(678, 'author', 'Бастаков В.А.')

# itas[itas.title == 'НКРЯ: о некоторых различиях в описании морфологии разных частей корпуса и об автоматическом построении синтаксических структур с учетом заранее известных морфологических характеристик ']
# Третий автор Виктор Сизов
itas = itas.set_value(1114, 'author', 'Сизов В.')

# itas[itas.title == 'Использование критерия Отсу в пространстве геометрических разрезов на примере задачи поиска линии горизонта ']
# Третий автор Дарья Крохина
itas = itas.set_value(1199, 'author', 'Крохина Д.')

# itas[itas.title == 'Reasoning over Knowledge-based Generation of Situations in Context Spaces to Reduce Food Waste ']
# Четвертый автор Sylvain Kubler
itas = itas.set_value(1964, 'author', 'Sylvain Kubler')

itas.head()

## Process authors -> split into names and surnames

def split_author_names(author):
    
    name = author.split()
    
    if len(name) == 1:
        first_name = author
        second_name = author
        
        return first_name, second_name
    elif len(name) == 3:
        first_name = author
        second_name = author
        
        return first_name, second_name
   
    # if len(name == 2)
    
    name1 = name[0]
    name2 = name[1]

    if '.' in name1:
        first_name = name1
        second_name = name2
    elif '.' in name2:
        second_name = name1
        first_name = name2
    else:
        first_name = name1
        second_name = name2
        
    return first_name, second_name

# for idx, val in enumerate(itas.author.values):
# #     if len(val.split(' ')) != 2:
# #         print(idx, val, len(val.split(' ')))
#     print(val.split(' '))

first_name = []
second_name = []
for auth in itas.author.values:
    name1, name2 = split_author_names(auth)
    first_name.append(name1)
    second_name.append(name2)
    
itas['Name'] = first_name
itas['Surname'] = second_name

itas[['Name', 'Surname']].drop_duplicates().sort_values(by=['Surname', 'Name'])

itas.to_excel('itas_2010-2017_v2.xls')

itas.head()

itas[itas.title == 'Сдвиговое и полусдвиговое преобразование Хаффа: генерация быстрых вычислительных схем ']

itas_2 = pd.read_excel('itas_2010-2017_v3.xls')
# после ручной обработки

itas_2['author_id_new'] = itas_2.Name + ' ' + itas_2.Surname

itas_2 = itas_2.drop_duplicates()

itas_2.to_excel('data_rus_author_v4.xls')

# В данной версии содержатся все данные, (почти) все авторы приведены к русскому имени,
# к одинаковой форме (либо имя + фамилия либо инициалы + фамилия) .

a = itas_2.author_id_new.values

auth, counts = np.unique(a, return_counts=True)

order = np.argsort(counts)[::-1]
auth = auth[order]
counts = counts[order]

for a, c in zip(auth, counts):
    print(a,c)
