==> itas_analysis/notebooks/2_add_gender.ipynb <==
import pandas as pd
import numpy as np

# itas_data = pd.read_excel('itas_2010-2017_v4.xls')
# itas_data.author_id_new = itas_data.Surname + ' ' + itas_data.Name
# itas_data.sort_values(by=['year', 'author_id_new'], inplace=True)
# itas_data.index = np.arange(2507)
# itas_data.set_value(173, 'Name', 'Татьяна')
# itas_data.set_value(173, 'Surname', 'Фролова')
# itas_data.set_value(173, 'author_id_new', 'Фролова Татьяна')

# itas_data.set_value(59, 'Name', 'Елена')
# itas_data.set_value(59, 'Surname', 'Бучацкая')
# itas_data.set_value(59, 'author_id_new', 'Бучацкая Елена')

# itas_data.set_value(1370, 'Surname', 'Логачева')
# itas_data.set_value(1371, 'Surname', 'Логачева')
# itas_data.set_value(1931, 'Surname', 'Логачева')

# itas_data.set_value(1370, 'author_id_new', 'Логачева Мария')
# itas_data.set_value(1371, 'author_id_new', 'Логачева Мария')
# itas_data.set_value(1931, 'author_id_new', 'Логачева Мария')

# v5 sorted by year -> author_id
# author_id is now Surname + Name (instead of Name + Surname),
# add gender
# fix Фролова Татьяна, Бучацкая Елена, Логачёва -> Логачева,
# fix index column according to new order
## itas_data.to_excel('itas_2010-2017_v5.xls')

itas_data = pd.read_excel('itas_2010-2017_v5.xls')
itas_data.head(50)

Добавить: пол, тематика автора, аффиляции

# sorted(itas_data.Surname.unique())

First check Name if return np.nan check Surname

name = name.strip()
if '.' not in name:
    if any 'йцукенгшщзхъфывапролджэячсмитьбю' in name:
        if end_with('a', 'я') -> female
        exceptions : ['Алипаша', 'Данила', 'Джошуа', 'Илья', 'Никита']
        else -> male
        exceptions : ['Аягоз']
    else:
        -> np.nan
else:
    -> np.nan
    
surname = surname.strip()

if any 'йцукенгшщзхъфывапролджэячсмитьбю' in surname:
    if end_with('a', 'я') -> female
    exceptions : ['Ахуджа', 'Гайдамака', 'Дюба', 'Заика',
                  'Каледа', 'Калита', 'Катруца', 'Кикавада',
                  'Ксаба', 'Мажейка', 'Мостафа','Притула','Цепа','Ша','Шарануца',
                   'Шешкус','Ялда',]
    else -> male
    exceptions : ['Киселюс', 'Комеч', 'Корх', 'Крейдлин', 'Круглик',
                 'Крюк', 'Крячко', 'Лавыш', 'Лейн','Лик', 'Малько',
                 'Маляр', 'Матусевич','Мендель','Мовсесян','Молдован',
                 'Молибог','Мороз','Намиот','Пастушок','Пахл','Пилипчук',
                 'Пино','Поклад','Прун','Пэш','Птушенко','Рамзи','Раффлер',
                 'Ревенко','Ройтберг','Рябко','Рябых','Савчик','Саенко',
                 'Сергушевич','Стерлинг', 'Сухре', 'Тавакколках','Талис',
                  'Тамазян', 'Танас','Тетко','Ткачев', 'Фертиг','Фридман',
                  'Фришман', 'Фуденберг', 'Хамад', 'Хегази', 'Холод','Циммер',
                 'Чанг', 'Чаплин', 'Чевэн', 'Шинкарюк','Шмитт', 'Шнепс-Шнеппе',
                  'Эммерт-Стреиб',]
else:
    -> np.nan

    
    
all exceptions fill with np.nan

def check_gender(name, surname):
    import re
    
    name = name.strip()
    surname = surname.strip()
    gender = np.nan

    if '.' not in name:
        if re.findall('[йцукенгшщзхъфывапролджэячсмитьбю]', name) != []:
            last_sym = name[-1]
            if (last_sym == 'а') or (last_sym == 'я') and (name not in ['Алипаша', 'Данила', 'Джошуа', 'Илья', 'Никита']): 
                gender = 'Female'
                return gender
            elif name == 'Аягоз':
                gender = 'Female'
                return gender
            else:
                gender = 'Male'
                return gender

    if np.isnan(gender):
        if re.findall('[йцукенгшщзхъфывапролджэячсмитьбю]', surname) != []:
            last_sym = surname[-1]
            if (last_sym == 'а') or (last_sym == 'я'):
                if surname in ['Ахуджа', 'Гайдамака', 'Дюба', 'Заика',
                               'Каледа', 'Калита', 'Катруца', 'Кикавада',
                               'Ксаба', 'Мажейка', 'Мостафа','Притула',
                               'Цепа','Ша','Шарануца', 'Шешкус','Ялда',]:
                    gender = np.nan
                else:
                    gender = 'Female'
                    return gender
            elif surname in ['Киселюс', 'Комеч', 'Корх', 'Крейдлин', 'Круглик',
                             'Крюк', 'Крячко', 'Лавыш', 'Лейн','Лик', 'Малько',
                             'Маляр', 'Матусевич','Мендель','Мовсесян','Молдован',
                             'Молибог','Мороз','Намиот','Пастушок','Пахл','Пилипчук',
                             'Пино','Поклад','Прун','Пэш','Птушенко','Рамзи','Раффлер',
                             'Ревенко','Ройтберг','Рябко','Рябых','Савчик','Саенко',
                             'Сергушевич','Стерлинг', 'Сухре', 'Тавакколках','Талис',
                              'Тамазян', 'Танас','Тетко','Ткачев', 'Фертиг','Фридман',
                              'Фришман', 'Фуденберг', 'Хамад', 'Хегази', 'Холод','Циммер',
                             'Чанг', 'Чаплин', 'Чевэн', 'Шинкарюк','Шмитт', 'Шнепс-Шнеппе',
                              'Эммерт-Стреиб']:
                gender = np.nan
            else:
                gender = 'Male'
                return gender
            
    return gender

name = 'Пень.'
surname = 'Колода'

check_gender(name, surname)

gender = []
for item in itas_data.iterrows():
    name = item[1].Name
    surname = item[1].Surname
    g = check_gender(name, surname)
    gender.append(g)

itas_data['Gender'] = gender

itas_data.sort_values(by=['year', 'author_id_new'], inplace=True)
itas_data.index = np.arange(2507)

#itas_data

itas_data.to_excel('itas_2010-2017_v5.xls')

# v6
# impute NaN genders manually
# still ~90 NaNs

itas_data = pd.read_excel('itas_2010-2017_v6.xls')

itas_data.sort_values(by=['year', 'author_id_new'], inplace=True)
itas_data.index = np.arange(2507)

itas_data.sort_values(by='Gender')

itas_data.to_excel('itas_2010-2017_v6.xls')
