==> itas_analysis/notebooks/4_add_subject_area.ipynb <==
Разметим 
самых активных авторов
и размножим их тематику на все опубликованные работы
Работы оставшиеся без тематики доразметим вручную

import pandas as pd

itas_data =  pd.read_excel('itas_2010-2017_v6.xls')

itas_data.groupby(by='author_id_new').count().sort_values('Name', ascending=False).index.values[:40]

itas_data.groupby(by='author_id_new').count().sort_values('Name', ascending=False)

itas_data[itas_data['author_id_new']=='Фролов Алексей']

author_subject_area = {
    'Хоров Евгений': 'telecom',
    'Базыкин Г.А.': 'biology',
    'Миронов Андрей': 'biology',
    'Николаев Дмитрий': 'computer vision',
    'Кондрашов А.С.':'biology',
    'Зяблов В.В.':'encryption',
    'Беляев Михаил': 'machine learning',
    'Бурнаев Евгений': 'machine learning',
    'Фаворов Александр':'biology',
    'Ляхов Андрей':'telecom',
    'Гельфанд Михаил': 'biology',
    'Ставровская Елена':'biology',
    'Красилов Артем':'telecom',
    'Панов Максим': 'machine learning',
    'Родионов Дмитрий':'biology',    
    'Кирьянов Антон':'telecom',
    'Приходько П.В.':'machine learning', 
    'Науменко С.А.':'biology',
    'Храмеева Екатерина':'biology',
    'Панчин Юрий':'biology',
    'Жилин И.В.':'encryption',
    'Рыбин Павел':'encryption',
    'Иванов Федор':'encryption',
    'Додонова Юлия':'machine learning',
    'Банков Дмитрий':'telecom',    
    'Сафонов Александр':'telecom',
    'Логачева Мария':'biology',
    'Миллер Борис':'computer vision',
    'Солопова Ирина':'biology',
    'Фролов Алексей':'encryption'
}

itas_data['Area'] = itas_data.author_id_new.map(author_subject_area)

title = itas_data[['title', 'Area']].dropna().title.values
Area = itas_data[['title', 'Area']].dropna().Area.values

new_map = dict(zip(title, Area))

itas_data['Area'] = itas_data.title.map(new_map)

itas_data.shape

itas_data[['title', 'Area']].dropna().shape
