==> itas_analysis/notebooks/5_get_info _from_pdf_garbage.ipynb <==
0. Get pdfs from 2017
1. get affilations
2. get references

import pandas as pd

pdf_info = pd.read_excel('itas_pdf_url2010-2016.xls')

pdf_info.head(5)
# there is url for each title
# digits of each url are names of file stored in ../pdf/

path = '/home/anvar/clustering/itas_analysis/pdf/1569605193.pdf'
path = '/home/anvar/Desktop/old desktop/old desktop 5/NET_presentation(1).pdf'

path = '/home/anvar/Desktop/old desktop/old desktop 5/Essentsializm.pdf'
path = '/home/anvar/Desktop/old desktop/old desktop 5/1608.00607.pdf'
path = '/home/anvar/Desktop/old desktop/old desktop 5/hw1-5.pdf'

# Tika works, try pdftotext
# read 

https://stackoverflow.com/questions/34837707/extracting-text-from-a-pdf-file-using-python

https://medium.com/@rqaiserr/how-to-convert-pdfs-into-searchable-key-words-with-python-85aab86c544f

https://github.com/jalan/pdftotext

https://github.com/chrismattmann/tika-python

from tika import parser

import tika
from tika import parser
parsed = parser.from_file(path)#, xmlContent=True)


''.

def parse_pdf_itas(path):
    
    import tika
    from tika import parser
    
    def decode_line(line):
        return line.encode('latin1').decode('cp1251').strip()
    
    def decode_char(line):
        _line = []
        for char in line:
            try:
                c = char.encode('latin1').decode('cp1251')
                c.append(_line)
            except:
                c = ''
        return ''.join(_line)
    
    
    pdf = parser.from_file(path)
    text_to_decode = pdf['content']
    text_to_decode = text_to_decode.strip().split('\n')
    
    text = []
    n_exceptions = 0
    for line in text_to_decode:
        try:
            single_line = decode_line(line)
            #single_line = line.encode('latin1').decode('cp1251').strip()
        except:
            n_exceptions += 1
            single_line = decode_char(line)
    
        text.append(single_line)
        
    return text

pdf = parser.from_file(path)
text_to_decode = pdf['content']
text_to_decode = text_to_decode.split('\n')

type(pdf['content'])

t = parse_pdf_itas(path)

t[:150]

from nltk.tokenize import word_tokenize

path

path = '/home/anvar/clustering/itas_analysis/pdf/1569320295.pdf'
path = '/home/anvar/clustering/itas_analysis/pdf/1569319241.pdf'

raw = parser.from_file(path)
#print(raw['content'])

text = raw['content']

text.split('\n')[-50:]

print(text[:1677].encode('latin1').decode('cp1251').strip())

print(text[1679:5547].encode('latin1').decode('cp1251').strip())

print(text[5548:5579].encode('latin1').decode('cp1251').strip())

print(text[5580:6016].encode('latin1').decode('cp1251').strip())

5580 + 436

print(text[6018:8656].encode('latin1').decode('cp1251').strip())

print(text[8657:8694].encode('latin1').decode('cp1251').strip())

8657 + 37

print(text[8696:].encode('latin1').decode('cp1251').strip())

text[8690:8710].encode('utf-8')

len(text)

print(text[30000:].encode('latin1').decode('cp1251').strip())

print(text[28000:])

raw.keys()

raw['status']

# print(text)

word_tokenize(text, language='russian')

# path = '/home/anvar/clustering/itas_analysis/pdf/1569320295.pdf'

# raw = parser.from_file(path)
# print(raw['content'])

type(raw['content'])

# raw['content'].split('\n')

import pdftotext

# Load your PDF
with open(path, "rb") as f:
    pdf = pdftotext.PDF(f)

pdf

print(pdf[0].encode('latin1').decode('cp1251'))

pdf[-2]

import pdfminer

from pdfminer.pdfinterp import PDFResourceManager, PDFPageInterpreter
from pdfminer.converter import TextConverter
from pdfminer.layout import LAParams
from pdfminer.pdfpage import PDFPage
from io import StringIO

def convert_pdf_to_txt(path):
    rsrcmgr = PDFResourceManager()
    retstr = StringIO()
    codec = 'utf-8'
    #codec = 'unicode_escape'
    #codec = 'cp1251'
    laparams = LAParams()
    device = TextConverter(rsrcmgr, retstr, codec=codec, laparams=laparams)
    fp = open(path, 'rb')
    interpreter = PDFPageInterpreter(rsrcmgr, device)
    password = ""
    maxpages = 0
    caching = True
    pagenos=set()

    for page in PDFPage.get_pages(fp, pagenos, maxpages=maxpages, password=password,caching=caching, check_extractable=True):
        interpreter.process_page(page)

    text = retstr.getvalue()

    fp.close()
    device.close()
    retstr.close()
    return text

path = '/home/anvar/Desktop/1569319241-2.pdf'

t = convert_pdf_to_txt(path)

from html.parser import HTMLParser

# docker run -ti --rm -v /home/anvar/clustering/itas_analysis/pdf:/pdf bwits/pdf2htmlex pdf2htmlEX --zoom 1.3 1569319241.pdf

path = '/home/anvar/clustering/itas_analysis/pdf/1569319241.html'

parser = HTMLParser()

parser.feed(path)

parser.handle_data(path)

from bs4 import BeautifulSoup

import urllib

with open(path, 'rb') as f:
    a = f.read()

soup = BeautifulSoup(a, 'html.parser')

BeautifulSoup()

import PyPDF2

with open(path, 'r') as f:
    a = f.read()

path = '/home/anvar/clustering/itas_analysis/pdf/1569319241.pdf'


#pdf = PyPDF2.PdfFileReader(p)

with open(path, encoding='unicode-escape') as f:
    x = f.read()

p = pdf.getPage(3)

ttt = p.extractText()

ttt

bytes(ttt, )

bytes(ttt, 'utf-8').decode('ascii')

StringIO(bytes(pdf.getPage(5)).decode('ascii', 'ignore'))

from io import StringIO

# Read each line of the PDF
pdfContent = StringIO(getPDFContent(path).encode("ascii", "ignore"))
# for line in pdfContent:
#     print
