==> itas_analysis/notebooks/8_get_itas2009.ipynb <==
url = 'http://iitp.ru/upload/content/539/itas09_proceedings.pdf'

itas2009 = ['Открытые задачи в области mesh-сетей на базе протокола IEEE 802.11',
'Сафонов А. А.',
'Метрика маршрутизации для трафика, чувствительного к задержкам',
'Хоров Е.',
'Физическая модель интерференции прямых соединений:классификация и анализ возможных случаев',
'Красилов А. И.',
'Оценка производительности сверхвысокоскоростных беспроводных сетей IEEE 802.11ac (VHT)',
'Якимов М.',
'Анализ эффекта самоинтерференции потоков в mesh-сетях IEEE 80211s',
'Пустогаров И. А.',
'Модель протокола маршрутизации AODV в NS-3',
'Бойко П., Бучацкая Е.',
'Оценка времени обслуживания пакетов в беспроводной сети под управлением протокола IEEE 802.16',
'Ляхов А. И., Лукин Д. В.',
'Трафик реального времени в mesh сетях стандартов IEEE 802.11s и FLAME',
'Андреев К., Бойко П.',
'Передача видеоданных по беспроводной сети с централизованным управлением и коррелированными помехами',
'Сидоркин О. С.',
'Улучшение качества видеотрансляции с использованием ранжирования пакетов в потоке MPEG',
'Титов А.Ю., Никифоров С.А., Коноплев В.В.',
'Надежная многоадресная рассылка в беспроводной меш-сети.',
'Цыганова А.',
'Динамическое распределение спектра между вторичными пользователями в сетях когнитивного радио',
'Ошмарин Д. В.',
'Применение генетического программирования и верификации моделей для построения автоматов управления системами со сложным поведением',
'К.В. Егоров, Ф.Н. Царев',
'Задача оптимального управления с управляемыми коэффициентами высокой размерности. Методы последовательного улучшения',
'Батурина О.',
'Метод разделения движений для синтеза дискретных законов управления динамическими системами',
'Рассадин Ю. М.',
'Модель адаптивного поведения агентов в мультиагентной системе управления экологической безопасностью',
'Доронина Е.Г',
'Fluidics in control of flight vehicles',
'Kovrygin P.',
'Об эффекте самоподобия трафика сервера, предоставляющего данные различного объёма',
'Цитович И. И., Титов И.',
'Оценивание амплитуды установившихся колебаний в релейных системах',
'Кочетков С.А.',
'Геоинформационная технология исследования динамики взаимодействия геофизических пространственно-временных процессов',
'Дерендяев А.Б., Гитис В. Г., Соболев Г. А.',
'Неманипулируемые механизмы распределения ресурсов на графовых моделях организационных систем',
'Коргин Н. А.',
'Проблемы оценки параметра Херста потоков требований мультисервисных сетей',
'Цитович И. И., Ноздрачев М.',
'Применение различных моделей на основе когнитивных карт для анализа и моделирования сложных проблемных ситуаций',
'Авдеева З.К., Максимов В.И.',
'Метод нахождения состояний гонки в потоках, работающих на разделяемой памяти',
'Кудрин М.Ю., Прокопенко А.С., Тормасов А.Г.',
'Сигнально-кодовые конструкции на базе эквидистантных кодов в системе множественного доступа с каскадным кодированием.',
'Зяблов В. В., Осипов Д.С.',
'Исследование пропускной способности системы множественного асинхронного доступа с пороговым приемом',
'Грошев Ф.В., Осипов Д.С.',
'Две конструкции сверточных ПМПП-кодов',
'Зяблов В. В., Кондрашов К. А.',
'Сравнение корректирующей способности МПП-кодов с кодами-компонентами разной избыточности',
'Зяблов В. В., Фролов А.А.',
'Применение исправления стираний для кодирования МПП-кодов',
'Зяблов В. В., Рыбин П.С.',
'Декодирование обобщенных кодов с локализацией ошибок.',
'Зяблов В.В, Кобозева И. Г.',
'Исправление малого числа ошибок и стираний с использованием кодов Рида-Соломона',
'Давыдов А. А., Калимуллин Р.',
'Многопоточная реализация алгоритмов кодирования и декодирования кодов с малой плотностью проверок',
'Зигангиров Д.К., Крещук А. А.',
'Автоматическое обнаружение и выделение стабильных изменений в видеопоследовательностях',
'Кондратенков И.Б., Рубанов Л.И.',
'Impact of the DEM Resolution on the Outdoor RF Propagation Models',
'Boyko P., Fakhriev D., Musienko M.',
'Реализация и валидация моделей распространения радиосигналов',
'Андреев К., Бойко П., Мазо А., Фахриев Д.',
'Глубоко аннотированный корпус русских текстов как обучающий электронный ресурс',
'Шеманаева О. Ю.',
'Интерпретация и практическое применение текстовых данных о сочетаемости лексем (на материале прилагательных)',
'Фролова Т.',
'Определение автора книги статистическими методами',
'Орлов Ю. Н., Осминин К. П.',
'Построение онтологии для узкой предметной области',
'Тимошенко С.П.',
'Исследование алгоритмов построения деревьев зависимостей на основе машинного обучения',
'Казенников А.О.',
'Атрибуты модальности в UNL',
'Диконов В.',
'Компьютерная система обучения владению языком — опыт практического применения',
'Дяченко П. ',
'Структурная организация медицинских знаний при дифференциальной диагностике разных форм алалии',
'Попова С.М., Перемячкина С.Ю., Лукашевич И.П.',
'Мышечные пост-эффекты и поддержание равновесия у здоровых и больных с врожденным слабоумием',
'Талис В.Л., Капитонов М.А, Максимова Е.В.',
'Роль движений глаз в распознавании лица по фотографии',
'Огнивов В., Рожкова Г.',
'О механизмах подавления восприятия хемосигналов в вомероназальной системе домовой мыши под действием стресса',
'Вознесенская А.Е.',
'Офтальмологический комплекс для диагностики и лечения зрения',
'Голубцов К.В., Грошев Ф.В., Трунов В.Г., Айду Э. А.',
'Способы разделения каналов акустической связи у близких видов Pentatomidae',
'Шестаков Л.С.',
'Воздействие электромагнитного излучения GSM диапазона на кровь, надпочечники и ц.н.с. крыс',
'Дурнева А. ',
'Транскрипция генов синтеза пролина у бактерий родов Marinobacter, Pseudomonas и Shewanella регулируется белком семейства tetR',
'Лопатовская К.В., Зверков О.А., Селиверстов А.В., Любецкий В.А.',
'Capture and release of coding DNA: evolution of bacterial genes by shift of stop codons',
'Vakhrusheva A.A., Kazanov M.D., Mironov A.A., Bazykin G.A.',
'Predicting novel protein-small molecule interactions using molecular modeling techniques ',
' Kalinina O. V., Russell R. B.',  
'Codon usage bias: biological function or neutral marker?',
'Vinogradova S., Vinogradov D., Mironov A.A.',
'Нитрат-нитритное дыхание протеобактерий: исследование эволюции регуляторных взаимодействий методами биоинформатики',
'Равчеев Д. А.',
'Prediction and Comparative Analysis of Transcriptional Regulatory Regions in Drosophila Genomes',
'Nikulova A.A., Favorov A.V., Sutormin R.A., Mironov A.A.',
'Conservative regions of proteins evolve under stronger positive selection',
'Bazykin G. A.; Kondrashov A. S.',
'Коэволюция белков подсемейства FadR и их сайтов связывания',
'Суворова И. А., Равчеев Д. А.',
'Регуляция SOS-ответа Протеобактерий: анализ коэволюции белка LexA и сайтов его связывания',
'М. Э. Борисова, Равчеев Д. А.',
'Эволюция химерного белка: исследование AraR зависимой регуляции методами сравнительной геномики',
'Лейн С.А., Равчеев Д.А.',
'Кислород-зависимая регуляция азотфиксации у Альфапротеобактерий – исследование методами сравнительной геномики',
'Климова Е. Ю., Равчеев Д.А.',
'Нуклеотидные замены в альтернативно сплайсируемых генах Drosophila',
'Ермакова Е., Малько Д.',
'Поиск мотивов связывания транскрипционных регуляторов множественной лекарственной устойчивости у бактерий',
'Жаров И., Казаков А. ',
'Анализ последовательности генома макронуклеуса Euplotes crassus',
'Цой О., Виноградов Д., Заика А., Гельфанд М.C',
'Автоматная интерпретация фрагментов последовательности ДНК',
'Епифанов А.С.',
'Iterative Statistical Approach for Discovering Evolutionary Conserved Members of Regulon',
'Stavrovskaya E.D., Rodionov D.A., Mironov A.A., Dubchak I., Novichkov P. S.',
'Автоматический перенос аннотации между геномами близкородственных организмов',
'Малько Д.',
'Моделирование конкуренции РНК-полимераз: влияние нокаута сигма субъединиц и температуры на экспрессию генов',
'Зверков О.А., Селиверстов А.В., Рубанов Л.И. ,Любецкий В.А.',
'Реконструкция эволюции белковых семейств',
'Горбунов К.Ю., Любецкая Е.В.',
'Антифаговый иммунитет прокариот',
'Артамонова И.И.',
'CRISPR Polymorphism in natural populations or Escherichia Coli',
'Казнадзей А.Д.',
'Особенности транскрипции хромосомных перестроек в геноме человека',
'Храмеева Е., Мазин П.',
'Evolution of signal peptide appearance/disappearance in bacterial genomes',
'Bykova N.A., Mironov A.A.',
'Эволюция моделей альтернативного сплайсинга паралогичных генов человека',
'Панчин А. Ю., Артамонова И.И.',
'Исключительные последовательности в геномах различных организмов',
'Митрофанов С. И., Панчин А. Ю., Спирин С. А., Панчин Ю. В., Алексеевский А. В.',
'Species demography and evolution of alternative splicing sites',
'Leushkin E. V., Denisov S.V., Gelfand M. S, Popadin K. Yu., Bazykin G. A.',
'Свойства компенсаторных замен в белках',
'Калинина О., Анашкина А., Мирина А., Раменский В.',
'Orthologous genes in mammals: what’s about orthologous CpG islands?',
'Fridman M., Oparina N., Makeev V. ',
'Геномная энциклопедия метаболических путей утилизации сахаров полученная методами сравнительного анализа бактериальных геномов.',
'Родионов Д. А.',
'Parallel evolutionary trajectories at sites of multiple amino acid replacements in Drosophila evolution',
'Terekhanova N., Bazykin G. A.',
'High Frequency of Reversals in Evolution of Vertebrates and Insects',
'Naumenko S., Kondrashov A., Bazykin G. A.',
'Оценка вариабельности протоспейсеров CRISPR-систем фагов, специфичных к Xanthomonas oryzae',
'Москвин И.А., Артамонова И. И.',
'ЭВОЛЮЦИЯ САЙТОВ СПЛАЙСИНГА: изменения в отдельных позициях и в сайтах в целом',
'Денисов С., Фаворов А., Миронов А., Нуртдинов Р., Гельфанд М.',
'Developing the UEPHA-MS network database the tool for organizing and accessing the clinical and biological information for patients with multiple sclerosis',
'Nurtdinov R.',
'Evolution of mitochondrial genome size: large genomes in small mammals and small genomes in large mammals',
'Popadin K.',
'Higher cryptic variation in rate of transversions than transitions',
'Seplarskij V., Bazykin G. A.',
'Происходит ли падение приспособленности в результате накопления слабовредных мутаций в малочисленных видах млекопитающих?',
'M. Баранова, Попадьин К.',
'On the Monomial Automorphism Group of s p -Components in the q-ary Hamming Code',
'Gorkunov E.V.',
'Изменчивые ключи подписи',
'Федюкович В. Е.',
'Торическая топология',
'Бухштабер В.М., Панов Т.Е.',
'Финитная аппроксимируемость обобщенно-транзитивных симметричных модальных логик',
'Кудинов А., Шапировский И.',
'Субоптимальные последовательные правила проверки непараметрических гипотез о распределениях с экспоненциально убывающими хвостами',
'Цитович Ф. И.',
'О методиках бустинга',
'Бурнаев Е., Приходько П.',
'Минимаксная задача скорейшего обнаружения разладки для броуновского движения с экспоненциальной функцией штрафа',
'Бурнаев Е.',
'Адаптивная регуляризация в задаче аппроксимации многомерных зависимостей',
'Беляев М, Бурнаев Е.',
'Динамическое управление доступом и скоростью обслуживания при активных пользователях',
'Миллер А. Б.',
'Приводимые семейства Шлезингера и решения систем Гарнье',
'Гонцов Р. Р.',
'Изомонодромные слияния фуксовых особенностей и особенности уравнения Шлезингера',
'Вьюгин И. В.']

def make_table(itas_list):
    import numpy as np
    import pandas as pd
    
    itas = np.array(itas_list)
    
    n_papers = itas.shape[0]
    _titles = itas[list(range(0, n_papers, 2))]
    _authors = itas[list(range(1, n_papers, 2))]
    
    titles = []
    authors = []
    
    for title, author in zip(_titles, _authors):
        paper_authors = author.split(',')
        for _author in paper_authors:
            titles.append(title.strip())
            authors.append(_author.strip())
            
    return np.array(titles), np.array(authors)

titles, authors = make_table(itas2009)

import pandas as pd
import numpy as np

titles = titles.reshape(192,1)
year = np.array([2009]*192).reshape(192,1)
authors = authors.reshape(192,1)
name = np.empty((192,1))
surname = np.empty((192,1))
author_id = np.empty((192,1))
gender = np.empty((192,1))

authors[0][0]

surname = []
for auth in authors:
    a = auth[0].split(' ')
    surname.append(a[0])
    
surname = np.array(surname).reshape(192,1)

itas2009 = pd.DataFrame(data=np.concatenate([titles,
                                             year,
                                            authors,
                                            name,
                                            surname,
                                            author_id,
                                            gender], axis=1), columns=[['title','year','author',
                                                         'Name', 
                                                         'Surname',
                                                         'author_id_new',
                                                         'Gender']])

itas2009.to_excel('itas2009.xls')

itas2009

# ITAS 2008
