==> itas_analysis/notebooks/9_get_itas2007-2008.ipynb <==
def simple_parse(path):
    import tika
    from tika import parser
    pdf = parser.from_file(path)
    text_to_decode = pdf['content']
    text_to_decode = text_to_decode.strip().split('\n')
    
    return text_to_decode

path = "/home/anvar/work/itas08_proceedings.pdf"

itas2008 = simple_parse(path)

# for i,line in enumerate(itas2008):
#     print(i, line)

itas2008 = itas2008[253:828]

cleaned = []
for line in itas2008:
    if (line != '') and (line != ' '):
        cleaned.append(line)

itas2008 = cleaned

len(itas2008)

import re

cleaned = []
for line in itas2008:

    if '...' in line:
        _line = re.sub('[.]', '', line)
        cleaned.append(_line)
    elif '___' in line:
        pass
    else:
        cleaned.append(line)

len(cleaned)

itas2008 = cleaned

cleaned = []
for line in itas2008:
    if ('80216' not in line) and ('80211' not in line):
        result = re.sub('[0-9]', '', line)
        cleaned.append(result)
    else:
        cleaned.append(line)

itas2008=cleaned

itas2008

itas2008 = ['Эффективность обобщенного механизма переключения битовой скорости в протоколе IEEE 80211',
 'Зарубинский И., Якимов М. ',
 'Исследование свойств мультиплексирования потоков в мультисервисных сетях   ',
 'Цитович И. , Титов И. ',
 'Об интервальных моделях телетрафика мультисервисных сетей   ',
 'Цитович И., Догадченко Н. ',
 'Пропускная способность широкополосного беспроводного канала точка-точка в режиме насыщения   ',
 'Шпилев С.А. ',
 'Распределения времени присоединения устройств к сети ECMA- (WiMedia)   ',
 'Сафонов А., Хоров Е. ',
 'Эффективность метрик стоимости соединений стандарта ZigBee   ',
 'Баскаков С.С. ',
 'Система поллинга с опросом типа "звезда" и ограниченным обслуживанием   ',
 'Семенова О.В., Сергеев П.А. ',
 'Анализ характеристик многокилометрового широкополосного беспроводного канала точка-точка с различными функциями управления   ',
 'Фахриев Д.Н., Шпилев С.А. ',
 'Организация повторных передач в системах связи с ретрансляционными станциями   ',
 'Уханова А.С. ',
 'Модель изменения очереди пакетов в беспроводной сети под управлением протокола IEEE 80216',
 'Лукин Д.В. ',
 'Исследование методов расчета антенн с несколькими облучателями для беспроводных сетей 80211',
 'Камышников И.Г., Воробьев В.М. ',
 'Новый алгоритм выбора скорости передачи в сетях стандарта IEEE 80211',
 'Адов А.А., Михин Д.Ю., Черкасов М.В. ',
 'Моделирование и анализ протокола маршрутизации HWMP стандарта IEEE 80211s ',
 'Андреев К., Лаконцев Д., Мазо А. ',
 'Эффективность применения МIМО в беспроводной городской сети передачи данных   ',
 'Жилин М.C., Воробьев В.М. ',
 'Оптимизация механизма определения занятости среды для mesh-сетей стандарта IEEE 80211s  ',
 'Коваленко А.В., Лаконцев Д.В., Пешков А.И. ',
 'Система доказательства теорем с автоматической генерацией знаний   ',
 'Раутиаинен А., Стефанюк В.Л. ',
 'Размещение фотографий на холсте с определением коэффициентов масштабирования   ',
 'Егорова М.А., Сафонов И.В. ',
 'ГеоТайм II - новая геоинформационная технология исследования пространственно-временных процессов   ',
 'Дерендяев А.Б., Метриков П.А., Гитис В.Г. ',
 'Анализ эффективности применения методов интерполяции для доопределения законов функционирования сложных систем   ',
 'Епифанов А.С. ',
 'Решение задачи нахождения соответствующего эталона методом нейронного кортекса   ',
 'Сергийчук М.Н., Михайлов А.М. ',
 'Интерактивные учебные пособия на основе средств виртуальной реальности в CALS – технологиях   ',
 'Ромакин В.А., Балабанов А.В. ',
 'Задачи оптимизации и информационного обеспечения при управлении ДПТ   ',
 'Синюков А. В. ',
 'Построение стохастической модели прогнозирования состояния вычислительной сети как решение задачи классификации и снижения размерности   ',
 'Еремеев В.Б. ',
 'Проектирование структур данных для комплекса сбора оперативной информации о случаях заболеваний   ',
 'Сурпин В.П., Волчков Д.В. ',
 'Функциональная модель агента-предприятия в мультиагентной системе управления экологической безопасностью   ',
 'Еременко Ю.И., Доронина Е.Г. ',
 'Вычисление геометрических параметров поверхности при наличии детерминированных возмущений   ',
 'Кочетков С.А. ',
 'Оптимизация процесса центрифугирования путем определения текущей массы в барабане   ',
 'Ханина Н.Ю. ',
 'Разработка и исследование макета отказоустойчивого струйного акселерометра для летательных аппаратов   ',
 'Коврыгин П.В. ',
 'Методика конкурентного анализа телекоммуникационных компаний   ',
 'Сафарли И.И. ',
 'Допредельная реализация метода иерархии управлений в системе управления двухзвенным манипулятором   ',
 'Уткин А.В. ',
 'Введение платы за участие в механизмы активной экспертизы   ',
 'Бурков В.Н., Коргин Н.А. ',
 'Модель беспроводного канала связи в условиях крупномасштабного замирания и случайного разброса параметров приемопередатчиков   ',
 'Баскаков С.С. ',
 'Исправление стираний низкоплотностными кодами Галлагера   ',
 'Зяблов В.В., Рыбин П.С ',
 'Адаптивная подстройка параметров кодирования речевого сигнала в системах IP-телефонии   ',
 'Иманкулов Р.А. ',
 'Оценивание характеристик канала в системе множественного доступа, основанной на перестроении ортогональных частот   ',
 'Осипов Д.С. ',
 'Сравнительный анализ помехоустойчивости плетёных свёрточных (woven) и турбо кодов   ',
 'Чеклин А.Б., Черкасов М.В., Адов А.А. ',
 'Программный комплекс для исследования систем передачи данных   ',
 'Грошев Ф.В., Зигангиров Д.К. ',
 'Применение вэйвлет-преобразования для анализа спектров EXAFS: исследование причины появления и анализ артефактов на спектрограммах   ',
 'Павлов В.Н., Чукалина М.В. ',
 'Развитие системы построения семантического представления текста c использованием языка-посредника UNL на базе лингвистического процессора ЭТАП-   ',
 'Диконов В. ',
 'Использование конечных автоматов для морфологического анализа и синтеза на основе словарей системы ЭТАП   ',
 'Казенников А.О. ',
 'Извлечение информации о сочетаемости лексем из аннотированного корпуса текстов   ',
 'Дяченко П., Фролова Т. ',
 'Автоматическое уточнение правил в самозагрузочной системе извлечения информации из текстов на естественном языке   ',
 'Масленников М. ',
 'Сравнение качества работы синтаксического анализатора системы ЭТАП и статистического синтаксического анализатора MaltParser на материале текстов из корпуса СинТагРус   ',
 'Сизов В.Г., Валеев Д.Р., Крейдлин Л.Г. ',
 'Лексические функции и возможности оптимизации поиска информации в интернете (на материале параметрических слов)   ',
 'Тимошенко С.П. ',
 'Спектрально-временные характеристики акустических сигналов птиц   ',
 'Гладышев А.М. ',
 'Различные способы неспецифической активации шагательной активности в условиях разгрузки нижних конечностей у человека   ',
 'Солопова И.А., Селионов В.А. ',
 'Наследственные нарушения слуха: дифференцирование  случаев ранней глухоты и тугоухости   ',
 'Баталова Д., Божкова В. ',
 'Роль структурной организации ЭЭГ-информации при выявлении патогенетического механизма алалии   ',
 'Перемячкина С.Г., Попова С.М., Лукашевич И.П. ',
 'Использование записей движений глаз для анализа механизмов зрительной оценки размеров на примере иллюзии Мюллера-Лайера   ',
 'Огнивов В.В., Максимов П.В., Рожкова Г.И. ',
 'Движения глаз при восприятии неоднозначных изображений типа куба Неккера   ',
 'Желтова М.Ю., Максимов П.В., Бастаков В.А., Панютин А.К. ',
 'Компьютерная система для исследования лабильности зрительной системы   ',
 'Грошев Ф.В., Голубцов К.В., Трунов В.Г., Айду Э.А. ',
 'Positive selection and alternative splicing of human genes   ',
 'Ramensky V., Nurtdinov R., Neverov A., Mironov A., Gelfand M. ',
 'SDPfox - Web-сервер для анализа субстратной специфичности белков   ',
 'Мазин П., Калинина О., Рахманинова А., Миронов А., Гельфанд М. ',
 'Подтверждение участия давления отбора в поддержании стабильности CpG островов в геноме   ',
 'Перцовская И., Опарина Н., Виноградов Д., Миронов А., Фаворов А. ',
 'Comparative Analysis of Transcription Regulation Strategies of Bacteriophages T and RB   ',
 'Pugach K., Severinov K., Gelfand M., Ravcheev D. ',
 'Комплексная регуляция дыхания бактерий: исследование методами сравнительной геномики   ',
 'Равчеев Д.А. ',
 'A Markov Chain Monte Carlo Technique for Identification of Combinations of Allelic Variants Underlying Complex Diseases   ',
 'Favorov A., Sudomoina M., Favorova O., Parmigiani G., Michael F. Ochs ',
 'Analyzing Drug-target Interaction from a Comparative Perspective   ',
 'Kalinina O., Russell R. ',
 'Изучение консервативности межгенных участков ДНК прокариот   ',
 'Пятницкий М.А. ',
 'Эволюция под отбором: исследование паттерна замен в сайтах сплайсинга   ',
 'Денисов С., Миронов А., Гельфанд М. ',
 'Фиксация точечных мутаций кодирующих участках генов плодовых мушек рода Drosophila   ',
 'Ермакова Е.О., Малько Д.Б., Гельфанд М.С. ',
 'Adaptation to the loss of leader protein by cardiovirus: analysis of compensatory mutations   ',
 'Fominykh K., Lidsky P. ',
 'Изучение регуляции транскрипции в геномах рода Shewanella путем автоматического поиска регуляторных мотивов в рамках функциональных подсистем   ',
 'Ставровская Е.Д., Сиприано М., Родионов Д.А., Дубчак И.Л., Миронов А.А., Гельфанд М.Г. ',
 "Non-' CpG islands in the human genome: probable involvement in transcription regulation   ",
 'Medvedeva Y.A., Fridman M.V., Oparina N.J., Malko D.B., Ermakova E.O., Kulakovskiy I.V., Makeev V.J. ',
 'Изучение системы деградации этаноламина методами сравнительной геномики   ',
 'Цой О.В., Равчеев Д.А., Гельфанд М.С., Мушегян А. ',
 'Alternative splicing tends to involve phosphorylation sites   ',
 'Kurmangaliyev Y.Z., Gelfand M.S. ',
 'Polymorphism Due to Multiple Amino Acid Substitutions at a Codon Site within Ciona savignyi   ',
 'Donmez N., Bazykin G.A., Brudno M., Kondrashov A.S. ',
 'Эволюция недавно дуплицировавшихся паралогов генома человека   ',
 'Панчин А., Артамонова И.И. ',
 'CRISPR-системы в метагеномах   ',
 'Сорокин В.А., Артамонова И.И. ',
 'ДНК коды, основанные на весовом стебельном сходстве Хэмминга   ',
 'Дьячков А.Г., Воронина А.Н. ',
 'Поиск вторичных структур в интронах представителей рода Drosophila   ',
 'Храмеева Е.Е., Миронов А.А., Гельфанд М.С., Первушин Д.Д. ',
 'Эволюционный анализ генов семейства SSX   ',
 'Шустрова Е., Артамонова И. ',
 'Регуляция дыхания бактерий: транскриптомика и сравнительная геномика   ',
 'Цыганова М., Равчеев Д. ',
 'Использование сравнительной геномики для улучшения достоверности аннотации сигнальных пептидов   ',
 'Быкова Н.А., Сутормин Р.А. ',
 'Эволюционные изменения структуры новой некодирующей РНК человека   ',
 'Бениаминов А., Вестхоф Э., Крол А. ',
 'Исследование корреляций между аминокислотными последовательностями транскрипционных регуляторов и нуклеотидными последовательностями их сайтов связывания с ДНК   ',
 'Коростелев Ю.Д., Лайкова О.Н., Рахманинова А.Б. ',
 'Deleterious effect of mitochondrial nucleotide repeats in Chordata   ',
 'Popadin K., Mamirova L., Kuptsov A. ',
 'Реконструкция эволюции систем регуляции генов метаболизма гистидина в бактериях   ',
 'Витрещак А. ',
 'Предсказание сайтов регуляции рибосомальных генов в бактериях   ',
 'Петрова С., Витрещак А., Гельфанд М. ',
 'Построение HMM, основанной на правилах взаиморасположения сайтов связывания транскрипционных факторов, и ее применение для поиска корегулируемых генов в геномах рода Drosophila   ',
 'Никулова А.А., Сутормин Р.А., Миронов А.А., Фаворов А.В. ',
 'Функциональная классификация генов полных бактериальных геномов на основе базы данных Clusters of Orthologous Groups (COG)   ',
 'Казанов М. ',
 'Корреляции между аминокислотными последовательностями транскрипционных регуляторов суперсемейства CRP-FNR и нуклеотидными последовательностями их операторов   ',
 'Митева Д.С., Степанова В.В. ',
 'Эволюция альтернативного сплайсинга генов млекопитающих   ',
 'Нуртдинов Р.Н. ',
 'Genome-wide assessment of codon usage correlation   ',
 'Vinogradov D.V., Mironov A.A. ',
 'Кислород-зависимая регуляция азотфиксации: исследование методами сравнительной геномики   ',
 'Климова Е.Ю., Равчеев Д.А. ',
 'Моделирование вычислительных систем непрерывными уравнениями   ',
 'Захаров В., Тормасов А. ',
 'Моделирование стационарных входных потоков данных для стохастических моделей дискретных систем   ',
 'Шляева А. ',
 'Алгоритм прогнозирования функции распределения нестационарного временного ряда   ',
 'Осминин К.П. ',
 'Многоэтапные процедуры проверки статистических гипотез и субоптимальность   ',
 'Цитович Ф. ',
 'Игровые стратегии и адаптивное прогнозирование   ',
 'Вьюгин В.В., Румянцев Е.Л., Цфасман Ф.М. ',
 'Уравнение Пенлеве VI и дополнительные особенности фуксовых уравнений   ',
 'Гонцов Р.Р. ',
 'О минимаксной асимптотической оптимальности первого порядка в задаче наискорейшего обнаружения изменения интенсивности пуассоновского процесса ..  ',
 'Бурнаев Е.В. ',
 'Оптимальное нечеткое агрегирование информации в системах принятия решений ',
 'Алфимцев А.Н., Девятков В.В. ',
 'Об одном достаточном условии PSPACE-разрешимости транзитивных модальных логик   ',
 'Шапировский И. ',
 'Сложность алгоритмов проверки и поддержания степеней непротиворечивости алгебраических байесовских сетей   ',
 'Сироткин А.В. ',
 'Барьерно - мультипликативные алгоритмы для двойственных задач   ',
 'Гасникова Е. ',
 'Сдвиг инвариантов в алгебрах Пуассона алгебр Ли токов на прямой   ',
 'Рыбников Л.Г. ',
 'Задача одношагового управления   ',
 'Мещерякова О.В. ',
 'Применение идемпотентного принципа соответствия для решения задач теории систем массового обслуживания   ',
 'Жбанов С.А. ',
 'Байесовские и марковские сети: логико-вероятностный вывод в базах фрагментов знаний с неопределенностью   ',
 'Тулупьев А.Л., Сироткин А.В. ',
 'Деформация границы простых множеств под действием растягивающей последовательности отображений   ',
 'Комеч С. ',
 'Протокол демонстрации K-кратного вхождения строки   ',
 'Федюкович В.Е., Шарапов В.Г. ',
 'Локализация семантических блоков на растровых изображениях без предварительного обучения   ',
 'Кулешов С.В., Зайцева А.А. ',
 'Быстрый алгоритм совмещения изображений документов в произвольной геометрической модели   ',
 'Усилин С., Николаев Д., Постников В. ',
 'Использование методов анализа неподвижных изображений для поиска событий в видеопоследовательностях   ',
 'Кондратенков И.Б., Рубанов Л.И. ',
 'Обработка данных в системах создания виртуального окружения   ',
 'Торчигин А.В. ',
 'Расчет количества текстурных кластеров для сегментации изображений на основе Марковских случайных полей   ',
 'Патана Е.И. ',
 'Использование цифрового нейронного кортекса на примере распознавания образов   ',
 'Лазаренко А.В., Михайлов А.М. ',
 'Метод идентификации типа документа по структуре проекций его изображения на координатные оси   ',
 'Безматерных П., Николаев Д., Постников В. ',
 'Метод восстановления изображений, подверженных радиальной дисторсии   ',
 'Николаев Д., Гладилин С., Карпенко С. ']

import numpy as np

itas2008 = np.array(itas2008)

def make_table(itas_list):
    import numpy as np
    import pandas as pd
    
    itas = np.array(itas_list)
    
    n_papers = itas.shape[0]
    _titles = itas[list(range(0, n_papers, 2))]
    _authors = itas[list(range(1, n_papers, 2))]
    
    titles = []
    authors = []
    
    for title, author in zip(_titles, _authors):
        paper_authors = author.split(',')
        for _author in paper_authors:
            titles.append(title.strip())
            authors.append(_author.strip())
            
    return titles, authors

titles, authors = make_table(itas2008)

titles, authors = np.array(titles), np.array(authors)

titles = titles.reshape(234,1)
year = np.array([2008]*234).reshape(234,1)
authors = authors.reshape(234,1)
name = np.empty((234,1))
surname = np.empty((234,1))
author_id = np.empty((234,1))
gender = np.empty((234,1))

surname = []
for auth in authors:
    a = auth[0].split(' ')
    surname.append(a[0])
    
surname = np.array(surname).reshape(234,1)

import pandas as pd

itas2008 = pd.DataFrame(data=np.concatenate([titles,
                                             year,
                                            authors,
                                            name,
                                            surname,
                                            author_id,
                                            gender], axis=1), columns=[['title','year','author',
                                                         'Name', 
                                                         'Surname',
                                                         'author_id_new',
                                                         'Gender']])

itas2008

# TODO 2007 и полная обработка итого xls файла (author_id и пр.)
